API changes for language resolution, bundle matching.
Language resolution is defined here as taking a potentially-incomplete locale
object and finding the best, fully-expanded CLDR locale by adding likely
subtags to "fill in the blanks".

For example, the resolved locale for "en" is "en-Latn-US".  Note that the
resolved locale for "en-XY" is "en-Latn-XY" since resolution does not correct
bad field values, it only fills in the most likely values for missing fields.

Bundle matching is defined as, given a resolved CLDR locale, finding
the bundle that best satisfies that locale.  The bundle "en" is used
to satisfy the locales "en-US", "und-Latn", "en-XY", etc.