Language matching extensions.
Now returns a Match object, providing access to the distance.  This lets
you detect the strength of the match. If the default is matched, it
returns a match with the maximum distance (100).

Added methods to return lists of matches, sorted by increasing distance.
This provides a bit more flexibility when doing intermediate matches
that may later be used for future matches.